fixed font family on cards
