--- hamza-client/src/modules/store-v2/component/side-menu.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { Box } from '@chakra-ui/react';
-
-const SideMenu = () => {
-    return (
-        <Box
-            borderRadius={'20px'}
-            maxWidth={'348.74px'}
-            width={'100%'}
-            backgroundColor={'#121212'}
-        >
-            side-menu
-        </Box>
-    );
-};
-
-export default SideMenu;
